LayoutTests/fast/visual-viewport/ios/visual-viewport-dimensions-during-scroll-with-keyboard.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->

<html>
<head>
    <meta name="viewport" content="initial-scale=1.0">
    <style>
        .spacer-above {
            background-color: silver;
            height: 600px;
        }
        .spacer-below {
            background-color: silver;
            height: 4000px;
        }
        input {
            display: block;
            margin: 60px 30px;
        }
    </style>
    <script src="../../../resources/js-test.js"></script>
    <script src="../../../resources/ui-helper.js"></script>
    <script>
        var jsTestIsAsync = true;

        function getScrollDownUIScript()
        {
            return `(function() {
                uiController.immediateScrollToOffset(0, 300);
            })();`;
        }

        function getFocusInputUIScript(x, y)
        {
            return `(function() {

                uiController.didShowKeyboardCallback = function() {
                    uiController.doAfterNextStablePresentationUpdate(function() {
                        uiController.uiScriptComplete();
                    });
                };
                uiController.singleTapAtPoint(${x}, ${y}, function() {});
            })();`;
        }

        async function runTest()
        {
            if (!window.testRunner || !testRunner.runUIScript)
                return;
            
            await new Promise(resolve => testRunner.runUIScript(getScrollDownUIScript(), resolve));
            var rect = document.getElementById('input').getBoundingClientRect();
            await new Promise(resolve => testRunner.runUIScript(getFocusInputUIScript(rect.left, rect.top + document.scrollingElement.scrollTop), resolve));
            window.scrollTo({
                top: 1000,
                behavior: 'smooth',
            });

            numEqualOfEqualVisualViewportRects = 0;
            numIterations = 0;
            var currentVisualViewportRect = JSON.stringify(internals.visualViewportRect());
            while (window.scrollY != 1000) {
                await new Promise(resolve  => requestAnimationFrame(resolve));
                numIterations++;

                var newVisualViewportRect = JSON.stringify(internals.visualViewportRect());
                if (currentVisualViewportRect == newVisualViewportRect) {
                    numEqualOfEqualVisualViewportRects += 1;
                } else {
                    break;
                }
            }

            shouldNotBe("numEqualOfEqualVisualViewportRects", "numIterations");
            finishJSTest();
        }

        window.addEventListener('load', runTest, false);
    </script>
</head>
<body>
<p>Tests that the visual viewport rect is updated during a scroll with the keyboard up.</p>
<div class="spacer-above"></div>
<input id="input" type="text"/>
<div class="spacer-below"></div>
</body>
</html>
