LayoutTests/fast/visual-viewport/ios/visual-viewport-dimensions-during-scroll-with-keyboard-expected.txt

PASS numEqualOfEqualVisualViewportRects is not numIterations
PASS successfullyParsed is true

TEST COMPLETE
Tests that the visual viewport rect is updated during a scroll with the keyboard up.
